fix 113

+++ ndsi-engine/src/main/java/br/com/abril/nds/integracao/ems0113/processor/EMS0113MessageProcessor.java
@@ -51,8 +51,8 @@ public class EMS0113MessageProcessor extends AbstractRepository implements Messa
 			
 			if (null != descontoLogistica ) {
 								
-				descontoLogistica.setPercentualDesconto(input.getPercentDesconto());
-				descontoLogistica.setPercentualPrestacaoServico(input.getPercentPrestServico());
+				descontoLogistica.setPercentualDesconto(input.getPercentDesconto().floatValue());
+				descontoLogistica.setPercentualPrestacaoServico(input.getPercentPrestServico().floatValue());
 				descontoLogistica.setDataInicioVigencia(input.getDataInicioDesconto());
 				
 				getSession().merge(descontoLogistica);
@@ -61,8 +61,8 @@ public class EMS0113MessageProcessor extends AbstractRepository implements Messa
 				
 				descontoLogistica.setId(null);//auto increment
 				descontoLogistica.setTipoDesconto(input.getTipoDesconto());
-				descontoLogistica.setPercentualDesconto(input.getPercentDesconto());
-				descontoLogistica.setPercentualPrestacaoServico(input.getPercentPrestServico());
+				descontoLogistica.setPercentualDesconto(input.getPercentDesconto().floatValue());
+				descontoLogistica.setPercentualPrestacaoServico(input.getPercentPrestServico().floatValue());
 				descontoLogistica.setDataInicioVigencia(input.getDataInicioDesconto());
 				
 				getSession().persist(descontoLogistica);
